refactor: moved state logic into SocialMediaButton and renamed SocialsTrayWrapper to SocialsTray
